Django/requirements.txt
anyio==3.7.1
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
crispy-bootstrap5==0.7
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.4
django-crispy-forms==2.0
djangorestframework-simplejwt==5.3.0
dodgy==0.2.1
duplicity==0.8.21
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.9.0
flake8==5.0.4
flake8-polyfill==1.0.2
fqdn==1.5.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.32
html5lib==1.1
httpie==3.2.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.14.0
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==7.0.0
notebook_shim==0.2.3
numpy==1.25.1
oauthlib==3.2.0
olefile==0.46
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pep8-naming==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.0.0
pluggy==1.2.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prospector==1.10.2
protobuf==3.12.4
psutil==5.9.5
psycopg==3.1.10
psycopg-binary==3.1.10
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.17.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.4.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-decouple==3.8
python-json-logger==2.0.7
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
referencing==0.30.0
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==1.0.0
requirements-detector==1.2.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
SecretStorage==3.3.1
semver==3.0.1
Send2Trash==1.8.2
setoptconf-tmp==0.3.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4.1
sqlparse==0.4.4
ssh-import-id==5.11
stack-data==0.6.2
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.19.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
word2number==1.1
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
